refactor: split components & improve styling and layout
